The scanning and parsing of the entire file is completed with this checkin.
Basically A big memory is allocated which holds array of rec_dataptr_t and then
the memory hold the contents of the entire file. Then, a scan is done over the
memory to populate the record array with the record releted data from the
contents of the file. To verify this, prints has been added.