This is probably is the final optimization:
When writing the sorted array back to disk, there were two writes. This means
that there were two system calls happening for one record. To reduce the number
of write system calls, I made the rarray[i].data_ptr to point to the record
itself in the main memory (Remember that entire file is read into the main
memory at once). So this pointer will act as a location in the memory where the
entire record's information is sotred. This will help us to write() the entire
record at once using this pointer with only one system call (write()).